req2_llm_finetune/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Vertex AI backend (chỉ cần khi EMPATHY_MODE=vertex)
google-cloud-aiplatform>=1.50.0
google-auth>=2.28.0
